+++ packages/components/src/navigator/navigator-screen/index.ts
@@ -1,1 +1,1 @@
-export { default as NavigatorScreen } from './component';
+export { NavigatorScreen } from './component';